drivers: counter: mcp7940n: move register defs into the driver source
The MCP7940N register-layout structs and enums were only used by the
driver, so move them into rtc_mcp7940n.c. The public header now only
declares mcp7940n_rtc_set_time(), the sole way for an application to
set the time on this device.
